+++ src/components/Fitbit.jsx
@@ -1,5 +1,5 @@
 import { AuthProvider } from "react-oauth2-code-pkce";
-import { authConfig } from "./fitbit-config";
+import { authConfig } from "../fitbit-config";
 import Home from "./Home";
 
 function Fitbit() {
