+++ src/components/Insight.jsx
